Change Button Creation
Proof of Concept was old one. Move over to using sprites instead of rectangles. This allows usage of the p5play physics engine for overlaps and overall makes for a better time. The buttons can be accessed using an array. This array, however, is private and only accessible from within the userInterface Class itself.