I have seen the eyes of god, and they said to me:
"woe, Nick, woe. For you have gazed unto the sins of the occult. Thine soul is tainted and heavy with the burden of sin. Go. Wash yourself and cleanse thineself, for the power of JavaScript blacked your very essence."

Oh, and i made it so the UI now overlays everything :P